requested changes assert matches new tensorflow example, move tests to `batchnorm_test.ts`, and replace batchnorm with batchnorm3D where appropriate.
