+++ src/math/batchnorm_test.ts
@@ -18,14 +18,14 @@
 import * as test_util from '../test_util';
 import {MathTests} from '../test_util';
 
-import {Array1D, Array3D} from './ndarray';
+import {Array1D, Array2D, Array3D} from './ndarray';
 
 // math.batchNormalization3D
 {
   // TODO(nsthorat): Fix the precision for byte-packed batchnorm.
   const epsilon = 1e-1;
   const tests: MathTests = it => {
-    it('simple batchnorm, no offset or scale, 2x1x2', math => {
+    it('simple batchnorm3D, no offset or scale, 2x1x2', math => {
       const x = Array3D.new([2, 1, 2], new Float32Array([2, 100, 4, 400]));
       const mean = Array1D.new([1, 2]);
       const variance = Array1D.new([2, 3]);
@@ -52,7 +52,7 @@ import {Array1D, Array3D} from './ndarray';
       variance.dispose();
     });
 
-    it('simple batchnorm, no offset, 2x1x2', math => {
+    it('simple batchnorm3D, no offset, 2x1x2', math => {
       const x = Array3D.new([2, 1, 2], new Float32Array([2, 100, 4, 400]));
       const mean = Array1D.new([1, 2]);
       const variance = Array1D.new([2, 3]);
@@ -81,7 +81,7 @@ import {Array1D, Array3D} from './ndarray';
       scale.dispose();
     });
 
-    it('simple batchnorm, no scale, 2x1x2', math => {
+    it('simple batchnorm3D, no scale, 2x1x2', math => {
       const x = Array3D.new([2, 1, 2], new Float32Array([2, 100, 4, 400]));
       const mean = Array1D.new([1, 2]);
       const variance = Array1D.new([2, 3]);
@@ -114,7 +114,7 @@ import {Array1D, Array3D} from './ndarray';
       offset.dispose();
     });
 
-    it('simple batchnorm, 2x1x2', math => {
+    it('simple batchnorm3D, 2x1x2', math => {
       const x = Array3D.new([2, 1, 2], new Float32Array([2, 100, 4, 400]));
       const mean = Array1D.new([1, 2]);
       const variance = Array1D.new([2, 3]);
@@ -184,6 +184,170 @@ import {Array1D, Array3D} from './ndarray';
 
   test_util.describeMathCPU('batchNormalization3D', [tests]);
   test_util.describeMathGPU('batchNormalization3D', [tests], [
+    {'WEBGL_FLOAT_TEXTURE_ENABLED': true, 'WEBGL_VERSION': 1},
+    {'WEBGL_FLOAT_TEXTURE_ENABLED': true, 'WEBGL_VERSION': 2},
+    {'WEBGL_FLOAT_TEXTURE_ENABLED': false, 'WEBGL_VERSION': 1}
+  ]);
+}
+
+// math.batchNormalization2D
+{
+  // TODO(nsthorat): Fix the precision for byte-packed batchnorm.
+  const epsilon = 1e-1;
+  const tests: MathTests = it => {
+    it('simple batchnorm2D, no offset or scale, 2x2', math => {
+      const x = Array2D.new([2, 2], new Float32Array([2, 100, 4, 400]));
+      const mean = Array1D.new([1, 2]);
+      const variance = Array1D.new([2, 3]);
+      const varianceEpsilon = .001;
+
+      const result = math.batchNormalization2D(
+          x, mean, variance, varianceEpsilon, undefined, undefined);
+
+      test_util.expectArraysClose(
+          result.getValues(), new Float32Array([
+            (x.get(0, 0) - mean.get(0)) * 1 /
+                Math.sqrt(variance.get(0) + varianceEpsilon),
+            (x.get(0, 1) - mean.get(1)) * 1 /
+                Math.sqrt(variance.get(1) + varianceEpsilon),
+            (x.get(1, 0) - mean.get(0)) * 1 /
+                Math.sqrt(variance.get(0) + varianceEpsilon),
+            (x.get(1, 1) - mean.get(1)) * 1 /
+                Math.sqrt(variance.get(1) + varianceEpsilon)
+          ]),
+          epsilon);
+
+      x.dispose();
+      mean.dispose();
+      variance.dispose();
+    });
+    it('simple batchnorm2D, no offset, 2x2', math => {
+      const x = Array2D.new([2, 2], new Float32Array([2, 100, 4, 400]));
+      const mean = Array1D.new([1, 2]);
+      const variance = Array1D.new([2, 3]);
+      const scale = Array1D.new([4, 5]);
+      const varianceEpsilon = .001;
+
+      const result = math.batchNormalization2D(
+          x, mean, variance, varianceEpsilon, scale, undefined);
+
+      test_util.expectArraysClose(
+          result.getValues(), new Float32Array([
+            (x.get(0, 0) - mean.get(0)) * scale.get(0) /
+                Math.sqrt(variance.get(0) + varianceEpsilon),
+            (x.get(0, 1) - mean.get(1)) * scale.get(1) /
+                Math.sqrt(variance.get(1) + varianceEpsilon),
+            (x.get(1, 0) - mean.get(0)) * scale.get(0) /
+                Math.sqrt(variance.get(0) + varianceEpsilon),
+            (x.get(1, 1) - mean.get(1)) * scale.get(1) /
+                Math.sqrt(variance.get(1) + varianceEpsilon)
+          ]),
+          epsilon);
+
+      x.dispose();
+      mean.dispose();
+      variance.dispose();
+      scale.dispose();
+    });
+
+    it('simple batchnorm2D, no scale, 2x2', math => {
+      const x = Array2D.new([2, 2], new Float32Array([2, 100, 4, 400]));
+      const mean = Array1D.new([1, 2]);
+      const variance = Array1D.new([2, 3]);
+      const offset = Array1D.new([4, 5]);
+
+      const varianceEpsilon = .001;
+
+      const result = math.batchNormalization2D(
+          x, mean, variance, varianceEpsilon, undefined, offset);
+
+      test_util.expectArraysClose(
+          result.getValues(), new Float32Array([
+            offset.get(0) +
+                (x.get(0, 0) - mean.get(0)) * 1 /
+                    Math.sqrt(variance.get(0) + varianceEpsilon),
+            offset.get(1) +
+                (x.get(0, 1) - mean.get(1)) * 1 /
+                    Math.sqrt(variance.get(1) + varianceEpsilon),
+            offset.get(0) +
+                (x.get(1, 0) - mean.get(0)) * 1 /
+                    Math.sqrt(variance.get(0) + varianceEpsilon),
+            offset.get(1) +
+                (x.get(1, 1) - mean.get(1)) * 1 /
+                    Math.sqrt(variance.get(1) + varianceEpsilon)
+          ]),
+          epsilon);
+      x.dispose();
+      mean.dispose();
+      variance.dispose();
+      offset.dispose();
+    });
+
+    it('simple batchnorm2D, 2x2', math => {
+      const x = Array2D.new([2, 2], new Float32Array([2, 100, 4, 400]));
+      const mean = Array1D.new([1, 2]);
+      const variance = Array1D.new([2, 3]);
+      const offset = Array1D.new([3, 4]);
+      const scale = Array1D.new([4, 5]);
+
+      const varianceEpsilon = .001;
+
+      const result = math.batchNormalization2D(
+          x, mean, variance, varianceEpsilon, scale, offset);
+
+      test_util.expectArraysClose(
+          result.getValues(), new Float32Array([
+            offset.get(0) +
+                (x.get(0, 0) - mean.get(0)) * scale.get(0) /
+                    Math.sqrt(variance.get(0) + varianceEpsilon),
+            offset.get(1) +
+                (x.get(0, 1) - mean.get(1)) * scale.get(1) /
+                    Math.sqrt(variance.get(1) + varianceEpsilon),
+            offset.get(0) +
+                (x.get(1, 0) - mean.get(0)) * scale.get(0) /
+                    Math.sqrt(variance.get(0) + varianceEpsilon),
+            offset.get(1) +
+                (x.get(1, 1) - mean.get(1)) * scale.get(1) /
+                    Math.sqrt(variance.get(1) + varianceEpsilon)
+          ]),
+          epsilon);
+      x.dispose();
+      mean.dispose();
+      variance.dispose();
+      scale.dispose();
+      offset.dispose();
+    });
+
+    it('batchnorm2D matches tensorflow, 3x3', math => {
+      const x = Array2D.new(
+          [3, 3], new Float32Array([0.3136892 , 0.92389025, 0.594782,
+                                    0.05021042, 0.67545404, 0.93910035,
+                                    0.13277993, 0.96474269, 0.88608916]));
+      const mean = Array1D.new([0.19526312, 0.74857256, 0.45166398]);
+      const variance = Array1D.new([0.22963001, 0.61521992, 0.46623685]);
+      const offset = Array1D.new([0.43098484, 0.77712237, 0.47916298]);
+      const scale = Array1D.new([0.62186907, 0.85673736, 0.19201061]);
+      const varianceEpsilon = .001;
+
+      const result = math.batchNormalization2D(
+          x, mean, variance, varianceEpsilon, scale, offset);
+
+      test_util.expectArraysClose(
+          result.getValues(), new Float32Array([
+            0.58433646, 0.96846228, 0.51936529,
+            0.24315402, 0.69732157, 0.61608542,
+            0.35007446, 1.01304821, 0.60119441
+          ]));
+
+      x.dispose();
+      mean.dispose();
+      variance.dispose();
+      scale.dispose();
+      offset.dispose();
+    });
+  };
+  test_util.describeMathCPU('batchNormalization2D', [tests]);
+  test_util.describeMathGPU('batchNormalization2D', [tests], [
     {'WEBGL_FLOAT_TEXTURE_ENABLED': true, 'WEBGL_VERSION': 1},
     {'WEBGL_FLOAT_TEXTURE_ENABLED': true, 'WEBGL_VERSION': 2},
     {'WEBGL_FLOAT_TEXTURE_ENABLED': false, 'WEBGL_VERSION': 1}
